# docs/swagger.yaml
basePath: /
definitions:
...
  models.UserLogin:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  models.UserRegister:
    properties:
      email:
        type: string
      fullname:
        type: string
      password:
        minLength: 6
        type: string
      role:
        type: string
    required:
    - email
    - fullname
    - password
    type: object
  models.UserResponse:
    properties:
      createdAt:
        type: string
      email:
        type: string
      fullname:
        type: string
      id:
        type: integer
      role:
        type: string
      token:
        type: string
      updatedAt:
        type: string
    type: object
  response.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
info:
  contact: {}
  description: Dokumentasi REST API menggunakan Gin dan Swagger
  title: API Koda Shortlink Documentation
  version: "1.0"
paths:
...
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Login user dengan email dan password. Menghasilkan access token
        dan refresh token yang tersimpan di server.
      parameters:
      - description: Login payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: Returns user data, access token, and refresh token
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  properties:
                    refreshToken:
                      type: string
                    token:
                      type: string
                    user:
                      $ref: '#/definitions/models.UserResponse'
                  type: object
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Email or password incorrect
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Login user
      tags:
      - Auth
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      description: Logout user dengan menghapus refresh token di server
      parameters:
      - description: Logout request payload
        in: body
        name: request
        required: true
        schema:
          properties:
            refreshToken:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Refresh token required
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to logout
          schema:
            $ref: '#/definitions/response.Response'
      summary: Logout user
      tags:
      - Auth
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh access token menggunakan refresh token yang tersimpan di
        server. Token baru hanya diberikan jika refresh token valid dan aktif di sessions
        table.
      parameters:
      - description: Refresh Token payload
        in: body
        name: body
        required: true
        schema:
          properties:
            refreshToken:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Returns new access token
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  properties:
                    token:
                      type: string
                  type: object
              type: object
        "400":
          description: Refresh token required
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Invalid or expired refresh token
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to generate access token
          schema:
            $ref: '#/definitions/response.Response'
      summary: Refresh access token
      tags:
      - Auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with fullname, email, password, and optional
        role. Default role is 'user' if not provided.
      parameters:
      - description: User registration payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UserRegister'
      produces:
      - application/json
      responses:
        "201":
          description: Returns the created user data
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Email already registered
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Register a new user
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"